default-values-and-collections: Restructure `kts` module
