+++ kts/src/main/resources/META-INF/services/javax.script.ScriptEngineFactory
@@ -1,1 +1,1 @@
-ru.potatophobe.konfig.KonfigKtsEngineFactory
+ru.potatophobe.konfig.kts.KonfigKtsScriptEngineFactory
